Move etl.py up to main module
It's the main driver for our code. It does depend on i2b2 code
right now, but that won't be true forever. Let's move it to a more
stable location over the long term.

Moved i2b2.config up as well, since that's purely support code for
the main loop.

Also, install etl.py as a 'cumulus-etl' script along with the module.
