+++ cumulus/i2b2/__init__.py
@@ -3,5 +3,3 @@
 from . import schema
 from . import extract
 from . import transform
-from . import config
-from . import etl
